requirements.txt
requests
Markdown
beautifulsoup4
click
python-dotenv
pyperclip